Noop code tidying and commenting
A lot of variable renaming and stuff, purely just for my comfort while
looking at this in the last few days. This is finally, *actually* the
version of the file that was running at the Village Underground
show.

I've tried to make this a no-op change, though I don't have a
Sonicsphere in my hands at the moment to test so I could be
wrong... don't take this as totally certain.